Move ping driver action to internal API
